+++ src/room517/asm/interpreter/cmd/orl.java
@@ -6,7 +6,7 @@ package room517.asm.interpreter.cmd;
  */
 public class orl extends ope {
     @Override
-    int operate(int left, int right) {
+    long operate(long left, long right) {
         return left | right;
     }
