+++ src/room517/asm/interpreter/cmd/sub.java
@@ -7,7 +7,7 @@ package room517.asm.interpreter.cmd;
 public class sub extends ope {
 
     @Override
-    int operate(int left, int right) {
+    long operate(long left, long right) {
         return left - right;
     }
